refactor(Worker): extract the rendering code to a separate class
